Euklides done

+++ src/main/java/Fibonacci/Main.java
@@ -6,7 +6,7 @@ public class Main {
     public static void main(String[] args) {
         Scanner myScanner = new Scanner(System.in);
         int number;
-        int result;
+        long result;
 
         System.out.println();
         System.out.println("Podaj liczbe ciagu ");
@@ -15,7 +15,7 @@ public class Main {
         System.out.println("Wynik fibonacci to " + result);
     }
 
-    public static int fibonacci(int number) {
+    public static long fibonacci(int number) {
         if (number <= 1) return number;
         else return fibonacci(number-1) + fibonacci(number-2);
     }
